# chore(storybook): update to storybook v7 with vite & internally update React v17