--- a/packages/lumx-react/storybook/.gitignore
+++ b/packages/lumx-react/storybook/.gitignore
@@ -1,2 +1,0 @@
-.yarn
-storybook-static