--- a/packages/lumx-react/storybook/build
+++ b/packages/lumx-react/storybook/build
@@ -1,4 +1,0 @@
-#!/bin/sh
-
-yarn install
-yarn build "$@"