--- a/packages/lumx-react/storybook/start
+++ b/packages/lumx-react/storybook/start
@@ -1,4 +1,0 @@
-#!/bin/sh
-
-yarn install
-yarn start "$@"