--- a/packages/site-demo/package.json
+++ b/packages/site-demo/package.json
@@ -48,9 +48,9 @@
         "moment-range": "^4.0.2",
         "prism-react-renderer": "^0.1.5",
         "prismjs": "^1.25.0",
-        "react": "^16.13.0",
+        "react": "^17.0.2",
         "react-docgen-typescript": "^2.2.2",
-        "react-dom": "^16.13.0",
+        "react-dom": "^17.0.2",
         "react-helmet": "^6.1.0"
     },
     "devDependencies": {