# General summary

Updating **StoryBook to v7** using the **vite.js config** instead of webpack (to get faster compile time and no dependency overlapping with the other packages in this repo)

**React had to be updated to v17** (previously v16) internally to correctly compile SB. The required version of React for the consumer of the `@lumx/react` lib will stil be **React v16 at minimum** (above is allowed, see `peerDependencies`).

SB Deployed on chromatic here: https://5fbfb1d508c0520021560f10-itzixtouoo.chromatic.com/

| | Before | After |
| --- | ---  | ---  |
| SB startup time | 26s (SB UI + stories loaded) | 9s (SB UI) and then 19s (SB UI + stories loaded) |
| SB hot code reload time | 1-2s | Almost instant |
| SB build | 33s | 26s |

Some warnings appear on SB startup about postcss config being out of date. We would need to update postcss which is  impacting for lumx-core, the demo site and SB so we won't do this for now. Those warning do not appear to impact the compilation of our SCSS in SB.

TODO:
- [x] Re-install Storybook v7 from scratch using vite (instead of webpack)
- [x] Make sure storybook can still deploy on chromatic
- [x] Fix incompatibilities
  - Update React to v17 internally (not in the peer dependencies, no impact on lib consumers)
  - Fix material theme switcher (load raw CSS)  
- [x] Fix broken stories listed by chromatic
  - GridColumn (only one column shown)
  - CommentBlock avatar is transformed into a button when it shouldn't
  - Thumbnail stories have many errors
- [x] Fix moment locale not updating in SB (displaying english date instead of the french date requested)
- [x] (non regression test) Make sure the demo site can still deploy
- [x] (non regression test) Test a NPM prerelease & test it in the product to make sure the changes in deps did not have weird behavior