@gcornut 2023-04-28
Chromatic build success but it looks like SCSS/CSS build in SB is misconfigured. 
In the GridColumn story: https://5fbfb1d508c0520021560f10-rrpzwyzanq.chromatic.com/?path=/story/lumx-components-grid-column-gridcolumn--columns
The grid style is broken