@gcornut 2023-04-28
Testing release CI build site, release note and deploy chromatic:
https://github.com/lumapps/design-system/actions/runs/4831806535
https://github.com/lumapps/design-system/actions/runs/4831806529
https://github.com/lumapps/design-system/actions/runs/4831806527