@gcornut 2023-05-03
Fixed most stories errors changing the SB actions config: removing automatic generation of action callback on props starting with `on`, on some component (like Thumbnail), adding a callback has an impact on the DOM/style (we should address that some day)