@gcornut 2023-05-05
Import `fr` moment locale in SB config to fix DatePicker stories