@gcornut 2023-05-05
Chromatic UI test OK => No diff with stories rendered on SB v6 ✔️